fix(app-auth): make sure our activity gets exec
By using Application.android.once we have the problem, that another
activity returning data to our app in the very moment we're runnig the
authentication code could lead to the problem that we'll never execute
the createTokenExchangeRequest.

By using Application.android.{on,off} we're sure that
createTokenExchangeRequest gets always executed.